[WIP] Flatten AnnData loader
