+++ src/loaders/anndata-loader/index.js
@@ -1,0 +1,3 @@
+import AnnDataLoader from './AnnDataLoader';
+
+export default AnnDataLoader;
