--- src/loaders/anndata-loaders/index.js
@@ -1,5 +1,0 @@
-import CellSetsZarrLoader from './CellSetsZarrLoader';
-import CellsZarrLoader from './CellsZarrLoader';
-import MatrixZarrLoader from './MatrixZarrLoader';
-
-export default { CellSetsZarrLoader, CellsZarrLoader, MatrixZarrLoader };
